requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
build==0.10.0
CacheControl==0.13.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
crashtest==0.4.1
cron-descriptor==1.4.0
cryptography==41.0.4
distlib==0.3.7
Django==4.2.6
django-celery-beat==2.5.0
django-filter==23.3
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
dulwich==0.21.6
filelock==3.12.2
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
jsonschema==4.17.3
keyring==24.2.0
kombu==5.3.2
more-itertools==10.1.0
msgpack==1.0.7
packaging==23.2
pexpect==4.8.0
Pillow==10.0.1
pipenv==2023.10.20
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==2.15.1
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shell==1.0.1
shellingham==1.5.3
six==1.16.0
sqlparse==0.4.4
stripe==7.0.0
tomlkit==0.12.1
trove-classifiers==2023.9.19
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.8
zipp==3.17.0
